some compilation fix

+++ include/ngchol/datatypes.hpp
@@ -11,10 +11,10 @@
 namespace LIBCHOLESKY {
 
 //typedef    long long int              LongInt;
-typedef    int              LongInt;
+typedef    long long int              LongInt;
 typedef    int                   Int;
-typedef    int              Idx32;
-typedef    int              Idx64;
+typedef    uint32_t              Idx32;
+typedef    uint32_t              Idx64;
 //typedef    uint64_t              Idx64;
 typedef    double                Real;
 typedef    std::complex<double>  Complex; // Must use elemental form of complex
